logout with token

+++ src/main/java/com/example/edoc/repository/QuyenReponsitory.java
@@ -3,5 +3,5 @@ package com.example.edoc.repository;
 import com.example.edoc.entity.Quyen;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuyenReponsitory extends JpaRepository<Quyen,Integer> {
+public interface QuyenReponsitory extends JpaRepository<Quyen,Long> {
 }
